+++ src/lib/project.js
@@ -554,7 +554,7 @@ async function scaffoldNext(projectName) {
   images: {
     unoptimized: true,
   },
-
+  output: 'export',
   /* Used to serve the Next.js app from a subdirectory (the GitHub repo name) and 
    * assetPrefix is used to serve assets (JS, CSS, images, etc.) from that subdirectory 
    * when deployed to GitHub Pages. The assetPrefix needs to be added manually to any assets